requirements.txt
click==6.7
kafka-python==1.3.5
python-dotenv==0.7.1
